feat: dynamic metadata project page
